@@ -20,9 +20,9 @@
     "@types/draftjs-to-html": "0.8.1",
     "@types/jest": "27.5.1",
     "@types/node": "16.11.20",
-    "@types/react": "17.0.38",
+    "@types/react": "18.0.15",
     "@types/react-d3-graph": "2.6.3",
-    "@types/react-dom": "17.0.11",
+    "@types/react-dom": "18.0.6",
     "@types/xml-js": "1.0.0",
     "assert": "2.0.0",
     "axios": "0.27.2",
@@ -53,11 +53,11 @@
     "plotly.js": "2.12.1",
     "ramda": "0.28.0",
     "ramda-adjunct": "3.1.0",
-    "react": "17.0.2",
+    "react": "18.2.0",
     "react-beautiful-dnd": "13.1.0",
     "react-color": "^2.19.3",
     "react-d3-graph": "2.6.0",
-    "react-dom": "17.0.2",
+    "react-dom": "18.2.0",
     "react-dropzone": "12.1.0",
     "react-hook-form": "7.31.3",
     "react-i18next": "11.16.9",
