@@ -23,7 +23,7 @@ function TimeSeriesManagement() {
     <UsePromiseCond
       response={res}
       ifPending={() => <SimpleLoader />}
-      ifRejected={(error) => <div>{error}</div>}
+      ifRejected={(error) => <div>{error?.toString()}</div>}
       ifResolved={(data) => (
         <Form autoSubmit config={{ defaultValues: data }}>
           <Fields study={study} />
